annotate s3 settings
